Chinchirorin: Add little log displays for every player
And make sure that debug info does not get onto these even when in
debug mode by having two output GameLog objects, one of which does not
get the debug lines.

TODO: playerActive info etc. should probably not be on this little log
screen, but it currently is. Those log messages should thus be split up.