Chinchirorin: Remove OnPlayerLeft logic that wasn't working anyway
We seem to have logic that works well for oya leaving now, with
OnOwnerShipTransferred, as the GameObject owner leaving will cause a
transfer to a "random" person that will not have `iAmPlayer == oya`.

As for other players leaving: that still needs handling. Personally I
think just a timeout will be good.

Also perhaps implementing the idea where everyone except the oya has
to press join and actually do a bet, until they are considered
actually in?