Chinchirorin: Make op argument always explicit
Now all the mkop_* series of functions return the op as a return
value, and the functions taking the ops apart take it as an explicit
argument.

Broadcast now takes an op as an argument, and assigns it to arg0
before requesting serialization.

This in preparation for having a queue of outgoing messages to
serialize (so that we no longer run the risk of some outgoing command
getting overwritten before getting serialized all proper and nice).