Chinchirorin: Remove some now-unneccessary waits in state machine
Since now we have the outgoing queue of ops, we can remove a lot of
places where we had introduced pauses to make sure serialization
happened properly. A few are still left in, where it will improve the
experience, such as giving you a bit more time to look at the dice the
in bowl before rethrow or moving to the next player (TODO: the
implementation of this wait would perhaps be best done inside
_OyaStateMachine itself, however. Perhaps in STATE_PREPARE_THROW?)

A noteable change is that we had to introduce a new op for enabling
the "Start Round" button on the oya side. Previously this button had
been enabled in the state machine itself, but since it is now possible
to arrive at STATE_WAITINGFORROUNDSTART before OPCODE_OYAREPORT has
been executed, what actually happened was that the button was enabled
first, but then the execution of OPCODE_OYAREPORT happened after,
disabling the button again. So to solve this we hade
STATE_WAITINGFORROUNDSTART emit OPCODE_WAITINGFORSTART, so that the
button will be enabled eventually. It might also be nice to tell all
players that they are waiting on the oya to do something. (TODO: this
will all change a bit if we change how joining works, however)

Generally speaking, any changes to the table itself should be an
opcode, and not done directly in the state machine (barring things
like disabling buttons OnClick to prevent double clicks, but that
mostly happens in the client side at this time)