Chinchirorin: Rewrite STATE_BALANCE and STATE_OYAPAYOUT for opqueue
We can just use an inline for loop now, ever since we got the opqueue.
We can simply queue all (up to) 4 OPCODE_BALANCE ops in the queue now,
and wait for them to get serialized later. No longer any need to loop
on the entire state with currentPlayer (currentPlayer is still useful
for STATE_PREPARE_THROW and STATE_THROW, however)