More TODO comments yay
This is an idea I have, but it would be a mess to implement (the
easiest would be to send the playerID to the owner, so it can bounce
that back, but events make that impossible).
Properly implemented it would prohibit the potential bug of several
players in the instance becoming the same player in the game.