Chinchirorin: Start adding timeouts
Now we have a few timeouts. It's also possible to tell in the leave op
whether it was due to timeout or not, so we can adjust the message.

TODO: Although maybe we should just make a separate op to tell everyone that
someone timed out? oyachange also might need a timeout parameter