Chinchiron: Waiting text fixes 2
There were still some bugs left in, including some particularly bad
one that could cause the udonbehavior to crash.

We also had to send along playeractive in OPCODE_WAITINGFORROUNDSTART
so we could activate the waiting for round start text in a way its
visible to all players who is waiting.