DieGrabSphere: More accurate inside detection
The previous way of checking would always treat the insideBowlCollider
as if it was a box collider. We use a sphere collider in the game, so
that sometimes gave false positives as to being inside it.

The new way seems a bit expensive, however. A potential optimization
could be to hard-code the fact that it is a sphere collider and then
just check the distance from center of die to center of spehere
collider.