Chinchirorin: Fix corner case: Better leaves instance. Oya pressed "Leave"
We break out the check in OnOwnershipTransferred and apply it in a few
more places, most notably all timeouts, so we can handle more
unexpected situations. We alse make the helper function MakeTableEmpty
that tries its best to make the table empty at all cost (including
BroadcastImmediate which doesn't rely on Update always getting called,
which might be a factor in the later vket implementation of this when
we have _VketUpdate instead)

Hopefully things are somewhat more robust, although very possibly
quite ugly (in particular having to call CheckOwnerIsOya before the
timeout is very ugly IMO)

TODO: Perhaps implement some sort of timeout on ownership not changing
on an oyachange (how? perhaps SendCustomEventDelayedSeconds that
checks that we are no longer the owner after a certain amount of
seconds. If we still are the owner, things are bad -> MakeTableEmpty)