Chinchirorin: Add more robustness around oyachange
We now add a paranoia function to check a few seconds after we issued
an oyachange. Of course if you are able to play an entire round within
this time range then we might actually unneccessarily reset the game.
But that seems very unlikely (a fix could be by doing something
similar to the other timeouts with also having a timeout time)