Chinchirorin: "Fix" a certain race condition on join
We now only set iAmPlayer when the server has recognized the join as
valid, giving the server the power to reject invalid joiners.

By adding a random nonce, that the server should repeat to us, we can
know that the join ACK was for us specifically (with a 1 in 32 chance
the both users involved in the race get a nonce collission and don't
manage to detect it resulting in two players on the same spot).
Considering that this race is going to be exceedingly unlikely in
itself, we consider this to be good enough for now.

I think the only 100% surefire way to solve this would be to take
playerId:s, but the current layout is not set up for doing that.