Create optimized variant of DieGrabSphere: DieGrabSphere2
In general this one is good for all games that don't require individual
dice re-rolls and such.

We immediately switch to use DieGrabSphere2 in Chinchirorin and for
DieGrabSphereDebug (which for some reason seems to work now??)

TODO: Make an abstract base class so that we can seamlessly plug &
play different DieGrabSphere variants without having to modify the
source code.