Chinchirorin: Use english names instead of japanese romaji
For the yaku names and calling "oya" dealer instead.
It doesn't make that much sense to use japanese names in romaji, when
the game is originally chinese, I suppose?

We still call the dealer "oya" internally in the code, because it is
short and convenient, and I am too lazy to change it everywhere.