Use preprocessor to conditionally support vket rules
Hopefully using VITDECK_HIDE_MENUITEM as a sign of vket tools will
work...

In most places we have just fixing the namespace, and there are a lot
of functions etc. that do not follow the rules.

In some places, such as DieGrabSphere2, I have started adding code to
use the vket alternative functions, like _VketFixedUpdate etc.