Die: Make it work under vket rules
The rigidbody member is made into a property that autofetches the
rigidbody and caches it.

Most importantly we ensure the die is always returned to isKinematic =
true. We could potentially add code to respawn the die in this case as
well (in case it has a VRCObjectSync at least), but relying on the
vket respawn logic seems to work good enough.

We are forced to make it non-synced, since we cant use continuous
sync (and we would need to use continuous sync if we use this
donbehaviour on the same object as there is an objectsync).

This only really matters for the edge case of partial re-rolls of a
group of dice, where the re-roll is someone else than the first
roller, however -- to transfer the roll results, essentially.

Though, thinking about it, it might be possible to simply rely on the
ObjectSync in such a case and redo the result calculation or so... oh
well, it doesn't matter for what I'm doing for now anyway (and since
I'm likely going to use DieGrabSphere2, we won't even need this
UdonBehaviour in the end)