Chinchirorin: Improve vket booth enter/exit handling
Unfortunately for now the stablest thing to do seems to treat everyone
outside the booth as completely unsynced. But at least this should
also cause the least load/trouble for all other people in the
instance.

One downside of this is that someone coming in to the booth cannot
join during the betting process, because you cannot sync during it.
To fix this we could: 1) always sync bets (fits in one ulong) 2)
separate out the logic that sets c_bets so that it always runs, even
when someone is outside the booth. Basically we ought to make the
client a bit more stateful.

A small ugly thing is that people getting into the booth during
STATE_WAITINGFORPLAYERS will also see a not so accurate timeout
display at the oya.

TODO: A thing that would improve the situation with the join buttons
      would be to instead of having a showbutton parameter sometimes as now,
      we have specific ops to set a clientside variable on/off.

TODO: Alternatively having the state variable be synced would simplify a lot of
      code w.r.t. when to show buttons. It would also be easier to
      know when synced etc. The current approach is to guess the state
      of the server based on what its last sent out message is, which
      is not really perfect and easy to break when you introduce more messages.