Chinchirorin: Add a "waiting for synchronization" text
Useful both in normal cases for late joiners, and for the vket case of
newcomers to the booth.

Although in the vket case we'd like to improve things so that someone
who isn't a late joiner, but is a newcomer to the booth can enter even
during the betting face. this should be possible in one of two ways:

  1) Have the bets in a synced variable (a single ulong ought to do).
  This would allow even late joiners to join during the betting phase.

  2) Do not treat outside of booth the same as being unsynced, and do
  some partial tasks in the case of being synced, but not in the
  booth, like updating the c_bets variable.