GameLog: Remove usage of Start for Vket rules
To do this we had to do the trick of using a property that checks if
the underlying variable has been initialized first and initializes it,
so that array gets created on first use.

We still have the _VketStart to make sure that clear gets run
eventually, but now only to make sure that there is no text on there.