Chinchirorin: Properly reset iAmPlayer and oya in _VketOnBoothExit
This caused a bug where if say P1 was oya. Then left the booth, then
someone else took over as oya and they were in the "waiting for
players" state, P1 would incorrectly see "Waiting for players..." and
"Leave" instead of seeing "Waiting for players..." at the actual oya,
and a "Join" button at their own place.

This could also potentially cause other bugs etc. etc.

In general it's a very good idea that we make sure that iAmPlayer gets
reset on booth exit, so no player keeps loosing or getting money even
when they are away from the booth.