Create a single player chinchirorin version
Needs to be done as a separate gimmick due to the how the regular one
is structured.

If we ever upgrade to UdonSharp 1.0 I hope that we will be able to get
rid of some of the duplication.

If I have time I would like to add a proper NPC character and stuff to
this to give it some more character.