[VKET]: Lock in on the Vket changes by commenting out macros
I do not know the exact environment that the final build of the vket
world will be in. Using VITDECK_HIDE_MENUITEM was a hack from start to
finish (but I'd love to see a define like this as an official feature
next vket around).

To avoid nasty surprises & build failures we simply comment out all
the conditionals, leaving behind the code that would have been there
in the case VITDECK_HIDE_MENUITEM was defined.

TODO: I think in the next commit I will also delete unused scripts and
prefabs etc. as a small cleanup/peace of mind task.