Work with VCC/U# 1.x
The little cute preprocessor trick we did to have vket compatibility
does not work for the namespaces any longer. It has still been kept
within the actual classes though, but that might be removed in a
future update.

Prefab fixes for U# 1.0 were also neccessary

Fixes for ChinchiroringSinglePlayer
We were using an implicit behaviour in old UdonSharp when a method
called with no arguments using the SendCustomEvent* series of
functions would call it using the previous argument variables to the
method. Effectively turning those into some sort of static variable.

We fix this by just making the ev and arg variables global variables,
as we need the ability to call the statemachine on a delay with
SendCustomEventDelayedSeconds. This is really what I should've done
from the start, but I was pressed for time when originally writing
thise and was satisfied with things simply working.