GameLog: add the _ToggleVisible function [vket2023summer]
I guess I added this specifically for vket last time to be able to
toggle the debug log, but then didn't put it in the master branch?

Anyway let's quickly and hackily reimplement it. I don't feel like
adding any additional parameters so just use the fact that the Text
object is the direct child of the Canvas in all my use cases.